Modified backend for easier frontend implementation
Created a UserSerializerPrivate which includes the email_verified field
This new serializer is now used with the LogInSerializer so JWT now contains username, user id and verification state.

Added a token refresh url as it was missing and is essential for JWT to work properly.

Added FRONTEND_URL to .env.dev

Changed email verification URL users have to visit in order to verify their email. It now uses frontend url from environment.
Fixed type for the company email