Refactored footer, user icon, deal view and all deals view
Footer now contains text and has a height of min-content
Removed user icon hover effects & cursor pointer.
Added height to Deal List View and Deal View if any of the API requests are still pending. This prevents footer from jumping to the top of the screen each time the view is loading.